Split prometheus api in service
